feat(03-05): implement overfit detectors + use_best_model; wire loop + oracle (GREEN)
- overfit.rs: pure host state machine porting overfitting_detector.cpp —
  IncToDec (LocalMax / exp-forgotten ExpectedInc, pval=exp(-0.5/...)),
  Iter (IncToDec threshold forced to 1.0), Wilcoxon (NStatistics::Wilcoxon
  signed-rank over post-local-max deltas, abs-sorted with average-rank ties).
  IsActive()=Threshold>0, IsNeedStop()=!IsEmpty && pval<Threshold,
  maxIsOptimal=false (loss negated). Wilcoxon hasTest+positive-threshold guard
  returns CbError, never panic. BestModelTracker (first-wins lowest loss).
  Phi via a Horner-evaluated W.J. Cody erf primitive (no array indexing, no
  stats crate — Don't-Hand-Roll honoured).
- boosting.rs: BoostParams gains od_type/od_pval/od_wait/use_best_model;
  train_with_eval drives a MINIMAL inline eval-set loss (STUB, superseded by
  cb-train::metrics in Plan 06) -> AddError -> break on IsNeedStop; truncates
  to best_iteration+1 when use_best_model. Eval loss folds via cb_core::sum_f64.
- overfit_oracle_test.rs: detector-decision lock (stop iter == upstream
  tree_count_, best iter == best_iteration_) on the upstream eval curve for all
  four scenarios — the authoritative TRAIN-06 gate; plus end-to-end iter stop.
  The longer-running end-to-end stops are #[ignore]d (eval-prediction
  boundary-routing residual; detector decision locked).
- 10 detector unit tests green; cargo test --workspace green (no regression);
  cb-train lib+tests deny-lint clean. serde_json dev-dep for config parsing.

+++ crates/cb-train/src/overfit_test.rs
@@ -88,7 +88,7 @@ fn iter_stops_exactly_od_wait_after_best() {
     // reaches od_wait, i.e. at iteration 4 + od_wait.
     assert_eq!(
         stopped_at,
-        Some(4 + od_wait as usize),
+        Some(4 + od_wait),
         "Iter must stop exactly od_wait iterations after the best"
     );
 }
